% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy_mean_ci.R
\name{accuracy_mean_ci}
\alias{accuracy_mean_ci}
\title{Calculate overall average accuracy and
the confidence interval of accuracy aggreating all Trees.
Will call Confusion_one_tree function to calculate confusion matrix for one tree.}
\usage{
accuracy_mean_ci(blrf, data, lower = 0.025, upper = 0.975)
}
\arguments{
\item{blrf}{object.}

\item{data}{data.frame object. Data to be predicted for accuracy.}

\item{lower}{numeric. Define lower bound of quantile.}

\item{upper}{numeric. Define upper bound of quantile.}
}
\value{
matrix. Average accuracy and confidence interval for each response variable.
}
\description{
Calculate overall average accuracy and
the confidence interval of accuracy aggreating all Trees.
Will call Confusion_one_tree function to calculate confusion matrix for one tree.
}
