% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_eval.R
\name{all_eval}
\alias{all_eval}
\title{Calculate overall average misclassfied number of Trees}
\usage{
all_eval(Trees)
}
\arguments{
\item{Trees}{list of tree object.}
}
\value{
numeric. Average misclassified number of Trees.
}
\description{
Calculate overall average misclassfied number of Trees
}
