% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blrf.R
\name{blrf}
\alias{blrf}
\title{blrf is used to fit Random Forest with Bag of Little Random Forests.
It can be used to carry out both regression and classification trees.}
\usage{
blrf(formula, data, gamma, b = NULL, s, r, n_var, split = "gini", core = 1)
}
\arguments{
\item{formula}{an object of class "formula".}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame) containing the variables in the model.}

\item{gamma}{numeric. User_defined sizing factor.}

\item{b}{numeric. An optional number of size of subsamples.}

\item{s}{numeric. number of subsamples.}

\item{r}{numeric. Number of trees.}

\item{n_var}{numeric. Number of variables to subset to build one tree.}

\item{split}{character string. Can be "deviance" or "gini". Default to be "gini".}

\item{core}{numeric. Must be positive. Number of core to use for parallel computing.
Default to 1, meaning no use of parallel computing.
If higher than 1, then implement the function with parallel computing with given number of cores.}
}
\value{
blrf object with different attributes given type of response variable.
}
\description{
blrf is used to fit Random Forest with Bag of Little Random Forests.
It can be used to carry out both regression and classification trees.
}
