% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confusion_one_tree.R
\name{Confusion_one_tree}
\alias{Confusion_one_tree}
\title{Calculate confusion matrix for each response variable for one tree.}
\usage{
Confusion_one_tree(one_tree, data)
}
\arguments{
\item{one_tree}{tree object.}

\item{data}{data.frame object.}
}
\value{
matrix. Confusion matarix values for each response variable for one tree.
}
\description{
Calculate confusion matrix for each response variable for one tree.
}
