% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval}
\alias{eval}
\title{Calculate misclassified number of one tree}
\usage{
eval(one_tree)
}
\arguments{
\item{one_tree}{tree object.}
}
\value{
numeric.
}
\description{
Calculate misclassified number of one tree
}
