% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_BLRF.R
\name{implement_BLRF}
\alias{implement_BLRF}
\title{Drive file to run BLRF}
\usage{
implement_BLRF(
  formula,
  data,
  gamma,
  b = NULL,
  s,
  r,
  n_var,
  split = "gini",
  core = 1
)
}
\arguments{
\item{formula}{A formula expression.}

\item{data}{data.frame. Original data.}

\item{gamma}{numeric. User_defined sizing factor.}

\item{b}{numeric. Size of subsamples.}

\item{s}{numeric. Number of subsamples.}

\item{r}{numeric. Number of trees.}

\item{n_var}{numeric. Number of variables to subset to build one tree.}

\item{split}{character string. Can be "deviance" or "gini". Default to be "gini".}

\item{core}{numeric. Must be positive. Number of core to use for parallel computing.
Default to 1, meaning no use of parallel computing.
If higher than 1, then implement the function with parallel computing with given number of cores.}
}
\value{

}
\description{
Drive file to run BLRF
}
