% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_tree.R
\name{one_tree}
\alias{one_tree}
\title{Implement random forest once}
\usage{
one_tree(formula, subsample, weights, n_var)
}
\arguments{
\item{formula}{A formula expression.}

\item{subsample}{data.frame.}

\item{weights}{list.}

\item{n_var}{numeric. Number of variables to sample.}
}
\value{
tree object.
}
\description{
Implement random forest once
}
