% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_tree_predict.R
\name{one_tree_predict}
\alias{one_tree_predict}
\title{Make prediction with one tree}
\usage{
one_tree_predict(one_tree, data)
}
\arguments{
\item{one_tree}{tree object.}

\item{data}{data.frame object. data to be predicted.}
}
\value{
data.frame object.
}
\description{
Make prediction with one tree
}
