% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.blrf.R
\name{predict.blrf}
\alias{predict.blrf}
\title{Make prediction with given data. User can indicate if output needs to be probability
for factor response or include confidence interval.
Default to output 95% confidence interval if include confidence interval.}
\usage{
\method{predict}{blrf}(
  blrf,
  newdata,
  confidence = F,
  probability = F,
  pretty = F,
  lower = 0.025,
  upper = 0.975
)
}
\arguments{
\item{blrf}{blrf object.}

\item{newdata}{data.frame. Data to be}

\item{confidence}{logical. If TRUE, then output will include confidence interval.}

\item{probability}{logical. If TRUE, then output will be predict probability for factor
type of blrf. If FALSE, then the output will be predict label for "factor"
type of blrf or predict value for "numeric" type of blrf.}

\item{pretty}{logical. If pretty is TRUE, then output character string output of ci for factor response,
not available for numeric response.}

\item{lower}{numeric. If confidence is TRUE, then define lower bound of ci.
Default to be 0.025.}

\item{upper}{numeric. If confidence is TRUE, then define upper bound of ci.
Default to be 0.975.}
}
\value{
list or matrix of prediction values (or with confidence interval).
}
\description{
Make prediction with given data. User can indicate if output needs to be probability
for factor response or include confidence interval.
Default to output 95% confidence interval if include confidence interval.
}
