% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ci_regression.R
\name{predict_ci_regression}
\alias{predict_ci_regression}
\title{Get Percentile Confidence Intervals for values of Regression Prediction}
\usage{
predict_ci_regression(blrf, new_data, lower = 0.025, upper = 0.975)
}
\arguments{
\item{blrf}{fitted model object of class blrf.}

\item{new_data}{data frame containing the values at which predictions are required.}

\item{lower}{numeric; lower percentile of confidence intervals.}

\item{upper}{nemeric; upper percentile of confidence intervals.}
}
\value{
list of confidence intervals.
}
\description{
Get Percentile Confidence Intervals for values of Regression Prediction
}
