% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_Tree.R
\name{Prediction_Tree}
\alias{Prediction_Tree}
\title{Make prediction with list of trees with given type of calculating the prediction}
\usage{
Prediction_Tree(Trees, newdata, type = "label")
}
\arguments{
\item{Trees}{list of tree objects.}

\item{type}{character string. Define the method to aggreate prediction. default to be "average". Choice from "average" to "frequency".}

\item{data}{data.frame. data to be predicted.}
}
\value{
data.frame. Aggregated prediction.
}
\description{
Make prediction with list of trees with given type of calculating the prediction
}
