% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_tree_regression.R
\name{prediction_tree_regression}
\alias{prediction_tree_regression}
\title{Make prediction with list of trees for regression}
\usage{
prediction_tree_regression(Trees, newdata)
}
\arguments{
\item{Trees}{list of tree objects.}

\item{data}{data.frame. data to be predicted.}
}
\value{
vector. Aggregated prediction.
}
\description{
Make prediction with list of trees for regression
}
