% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_predict.R
\name{trees_predict}
\alias{trees_predict}
\title{Make prediction with list of trees with given type of calculating the prediction}
\usage{
trees_predict(Trees, data, type = "average")
}
\arguments{
\item{Trees}{list of tree objects.}

\item{data}{data.frame. data to be predicted.}

\item{type}{character string. Define the method to aggreate prediction. default to be "average". Choice from "average" to "frequency".}
}
\value{
data.frame. Aggregated prediction.
}
\description{
Make prediction with list of trees with given type of calculating the prediction
}
