% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implement_BLRF.R
\name{implement_BLRF}
\alias{implement_BLRF}
\title{Drive file to run BLRF}
\usage{
implement_BLRF(formula, data, gamma, b = NULL, s, r, n_var, core = 1)
}
\arguments{
\item{formula}{A formula expression.}

\item{data}{data.frame. Original data.}

\item{gamma}{numeric. User_defined sizing factor.}

\item{b}{numeric. Size of subsamples.}

\item{s}{numeric. Number of subsamples.}

\item{r}{numeric. Number of trees.}

\item{n_var}{numeric. Number of variables to subset to build one tree.}

\item{core}{numeric. Number of core.}
}
\value{

}
\description{
Drive file to run BLRF
}
