% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_tree.R
\name{prediction_tree}
\alias{prediction_tree}
\title{Make prediction with list of trees with given type of calculating the prediction}
\usage{
prediction_tree(Trees, newdata, type = "label")
}
\arguments{
\item{Trees}{list of tree objects.}

\item{type}{character string. Define the output to be probability matrix or vector of label, default to be "label". Choice from "label" to "probability".}

\item{data}{data.frame. data to be predicted.}
}
\value{
data.frame. Aggregated prediction.
}
\description{
Make prediction with list of trees with given type of calculating the prediction
}
